# contracts/openapi/public-api.yaml
openapi: 3.1.0
...
info:
  title: EduTwin Public API
  version: 2.0.0
  license:
    name: Course demonstration use
    identifier: LicenseRef-Course-Demonstration
  description: |
    EduTwin browser-facing contract. MySQL is the system of record. Redis values,
    task state, and server-sent events are recoverable projections. Every traced
    response identifies the accepted answer event, analysis job, preallocated or
    persisted snapshot, data versions, and requested and effective model versions.
    Authentication uses a stateless JWT with one active role and all granted roles.
    Role, password, and account changes advance the token version and invalidate older
    tokens. Logout consists of discarding the JWT in the browser.
servers:
  - url: http://127.0.0.1:8080
    description: Local Compose deployment
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/analysis/jobs/{jobId}:
    parameters:
      - $ref: '#/components/parameters/JobId'
    get:
      tags: [Analysis Jobs]
      operationId: getAnalysisJob
      summary: Return durable analysis job state
      description: Students may access their jobs; teachers may access jobs in assigned courses.
      x-rbac: [STUDENT_SELF, TEACHER_ASSIGNED_TO_COURSE]
      responses:
        '200':
          description: Current durable job state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/analysis/jobs/{jobId}/events:
    parameters:
      - $ref: '#/components/parameters/JobId'
      - $ref: '#/components/parameters/LastEventId'
    get:
      tags: [Analysis Jobs]
      operationId: streamAnalysisJob
      summary: Stream durable analysis progress
      description: |
        Each event uses the persisted per-job sequence as the SSE id. A reconnect
        with Last-Event-ID replays later durable events and then continues live.
        Authorization is identical to the job resource.
      x-rbac: [STUDENT_SELF, TEACHER_ASSIGNED_TO_COURSE]
      responses:
        '200':
          description: Server-sent job events ending in job.completed or job.failed.
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-cache
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/SseJobEvent'
              x-sse-data-only: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/admin/users/{userId}/organization:
    parameters:
      - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Administration]
      operationId: replaceAdminUserOrganization
      summary: Replace the user's primary organization membership
      x-rbac: [ADMIN]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UserOrganizationReplaceRequest' } } }
      responses:
        '204': { description: User organization replaced. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/organizations:
    get:
      tags: [Administration]
      operationId: listAdminOrganizations
      summary: Return the enabled and disabled organization hierarchy
      x-rbac: [ADMIN]
      responses:
        '200': { description: Organization hierarchy., content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationList' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Administration]
      operationId: createAdminOrganization
      summary: Create an organization unit
      x-rbac: [ADMIN]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationWriteRequest' } } }
      responses:
        '201': { description: Organization created., content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationUnit' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/admin/organizations/{organizationId}:
    parameters:
      - { name: organizationId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Administration]
      operationId: updateAdminOrganization
      summary: Update, move, enable, or disable an organization unit
      x-rbac: [ADMIN]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationWriteRequest' } } }
      responses:
        '200': { description: Organization updated., content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationUnit' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/admin/organizations/{organizationId}/references:
    parameters:
      - { name: organizationId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Administration]
      operationId: getAdminOrganizationReferences
      summary: Preview user and course references before disabling an organization
      x-rbac: [ADMIN]
      responses:
        '200': { description: Organization reference counts., content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationReferences' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/academic-terms:
    get:
      tags: [Administration]
      operationId: listAdminAcademicTerms
      summary: List academic terms and lifecycle states
      x-rbac: [ADMIN]
      responses:
        '200': { description: Academic terms., content: { application/json: { schema: { $ref: '#/components/schemas/AcademicTermList' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Administration]
      operationId: createAdminAcademicTerm
      summary: Create an academic term
      x-rbac: [ADMIN]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AcademicTermWriteRequest' } } }
      responses:
        '201': { description: Academic term created., content: { application/json: { schema: { $ref: '#/components/schemas/AcademicTerm' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/admin/academic-terms/{termId}:
    parameters:
      - { name: termId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Administration]
      operationId: updateAdminAcademicTerm
      summary: Update or enable an academic term
      x-rbac: [ADMIN]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AcademicTermWriteRequest' } } }
      responses:
        '200': { description: Academic term updated., content: { application/json: { schema: { $ref: '#/components/schemas/AcademicTerm' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/counselor/students:
    get:
      tags: [Counselor]
      operationId: listCounselorStudents
      summary: List students in the counselor's assigned academic groups
      x-rbac: [COUNSELOR]
      parameters:
        - name: query
          in: query
          schema: { type: string, maxLength: 100 }
        - name: cohortYear
          in: query
          schema: { type: integer, minimum: 2000, maximum: 2200 }
        - name: className
          in: query
          schema: { type: string, maxLength: 200 }
        - name: riskBand
          in: query
          schema: { $ref: '#/components/schemas/RiskBand' }
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        '200':
          description: Paged counselor student summaries.
          content: { application/json: { schema: { $ref: '#/components/schemas/CounselorStudentPage' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/counselor/students/{studentId}/overview:
    parameters:
      - $ref: '#/components/parameters/StudentId'
    get:
      tags: [Counselor]
      operationId: getCounselorStudentOverview
      summary: Return read-only cross-course statistics for an assigned student
      x-rbac: [COUNSELOR]
      responses:
        '200':
          description: Cross-course student overview.
          content: { application/json: { schema: { $ref: '#/components/schemas/CounselorStudentOverview' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/counselor/classes/compare:
    get:
      tags: [Counselor]
      operationId: compareCounselorClasses
      summary: Compare aggregate learning and risk indicators across assigned classes
      x-rbac: [COUNSELOR]
      parameters:
        - name: classNames
          in: query
          required: true
          schema: { type: array, minItems: 1, uniqueItems: true, items: { type: string, minLength: 1, maxLength: 200 } }
          style: form
          explode: true
        - $ref: '#/components/parameters/AnalyticsPeriod'
      responses:
        '200': { description: Class comparison., content: { application/json: { schema: { $ref: '#/components/schemas/CounselorClassComparisonResult' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Stateless JWT containing one role plus userId and authorization-version claims; logout discards the token client-side.
  parameters:
...
    StudentId:
      name: studentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    LastEventId:
      name: Last-Event-ID
      in: header
      required: false
      schema:
        type: string
        pattern: '^[0-9]+$'
...
    AnalyticsPeriod:
      name: period
      in: query
      required: false
      schema:
        type: string
        enum: [7D, 30D, TERM]
        default: 30D
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Authenticated principal lacks the required role or scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Resource not found within the authorized scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: Idempotency conflict, stale snapshot, or immutable-version conflict.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    TraceRef:
      type: object
      additionalProperties: false
      required:
        - correlationId
        - answerEventId
        - analysisJobId
        - snapshotId
        - dataVersions
        - requestedModelVersions
        - effectiveModelVersions
      properties:
        correlationId:
          type: string
          format: uuid
        answerEventId:
          type: string
          format: uuid
        analysisJobId:
          type: string
          format: uuid
        snapshotId:
          type: string
          format: uuid
          description: Preallocated for QUEUED jobs and persisted unchanged on completion.
        dataVersions:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/DataVersionRef'
        requestedModelVersions:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ModelVersionRef'
        effectiveModelVersions:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ModelVersionRef'
...
    RiskBand:
      type: string
      enum: [LOW, MEDIUM, HIGH]
...
    AnalysisJobStatus:
      type: string
      enum: [QUEUED, PROCESSING, COMPLETED, FAILED]
    AnalysisStage:
      type: string
      enum: [QUEUED, MODEL_INFERENCE, SNAPSHOT_PERSISTENCE, PLAN_GENERATION, EVENT_PUBLICATION, COMPLETED, FAILED]
    AnalysisResultLinks:
      type: object
      additionalProperties: false
      required: [twinUrl, learningPlanUrl, diagnosisUrl]
      properties:
        twinUrl:
          type: [string, 'null']
        learningPlanUrl:
          type: [string, 'null']
        diagnosisUrl:
          type: [string, 'null']
    AnalysisFailure:
      type: object
      additionalProperties: false
      required: [errorCode, message, retryable, failedAt]
      properties:
        errorCode:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        retryable:
          type: boolean
        failedAt:
          type: string
          format: date-time
    AnalysisJob:
      type: object
      additionalProperties: false
      required:
        - jobId
        - studentId
        - courseId
        - status
        - stage
        - lastEventSequence
        - submittedAt
        - startedAt
        - completedAt
        - failure
        - result
      properties:
        trace:
          $ref: '#/components/schemas/TraceRef'
          x-edutwin-visibility: internal
        jobId:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/AnalysisJobStatus'
        stage:
          $ref: '#/components/schemas/AnalysisStage'
        lastEventSequence:
          type: integer
          format: int64
          minimum: 0
          description: Highest persisted analysis_job_event sequence visible to SSE recovery.
        submittedAt:
          type: string
          format: date-time
        startedAt:
          type: [string, 'null']
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
        degraded:
          type: boolean
          x-edutwin-visibility: internal
        degradationReasons:
          type: array
          uniqueItems: true
          x-edutwin-visibility: internal
          items:
            type: string
            enum:
              - KNOWLEDGE_MODEL_UNAVAILABLE
              - RISK_MODEL_UNAVAILABLE
              - MODEL_TIMEOUT
              - MODEL_CONTRACT_ERROR
              - DEEPSEEK_UNAVAILABLE
              - DEEPSEEK_TIMEOUT
              - LLM_UNAVAILABLE
              - LLM_TIMEOUT
              - REDIS_RETRY
        failure:
          oneOf:
            - $ref: '#/components/schemas/AnalysisFailure'
            - type: 'null'
        result:
          $ref: '#/components/schemas/AnalysisResultLinks'
    SseJobEvent:
      type: object
      additionalProperties: false
      required: [sequence, eventType, occurredAt, job]
      properties:
        sequence:
          type: integer
          format: int64
          minimum: 1
        eventType:
          type: string
          enum:
            - job.queued
            - job.processing
            - predictions.computed
            - twin.snapshot-created
            - learning-plan.created
            - job.completed
            - job.failed
            - heartbeat
        occurredAt:
          type: string
          format: date-time
        job:
          $ref: '#/components/schemas/AnalysisJob'
...
    OrganizationUnit:
      type: object
      additionalProperties: false
      required: [organizationId, code, displayName, unitType, parentId, enabled, children]
      properties:
        organizationId: { type: string, format: uuid }
        code: { type: string }
        displayName: { type: string }
        unitType: { type: string, enum: [COLLEGE, DEPARTMENT, MAJOR, CLASS] }
        parentId: { type: [string, 'null'], format: uuid }
        enabled: { type: boolean }
        children: { type: array, items: { $ref: '#/components/schemas/OrganizationUnit' } }
    OrganizationList:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/OrganizationUnit' } }
        total: { type: integer, minimum: 0 }
    OrganizationWriteRequest:
      type: object
      additionalProperties: false
      required: [code, displayName, unitType]
      properties:
        code: { type: string, minLength: 1, maxLength: 80 }
        displayName: { type: string, minLength: 1, maxLength: 160 }
        unitType: { type: string, enum: [COLLEGE, DEPARTMENT, MAJOR, CLASS] }
        parentId: { type: [string, 'null'], format: uuid }
        enabled: { type: boolean }
    OrganizationReferences:
      type: object
      additionalProperties: false
      required: [users, courses]
      properties:
        users: { type: integer, minimum: 0 }
        courses: { type: integer, minimum: 0 }
    UserOrganizationReplaceRequest:
      type: object
      additionalProperties: false
      required: [organizationId]
      properties:
        organizationId: { type: string, format: uuid }
    AcademicTerm:
      type: object
      additionalProperties: false
      required: [termId, code, displayName, startsOn, endsOn, enabled]
      properties:
        termId: { type: string, format: uuid }
        code: { type: string }
        displayName: { type: string }
        startsOn: { type: string, format: date }
        endsOn: { type: string, format: date }
        enabled: { type: boolean }
    AcademicTermWriteRequest:
      type: object
      additionalProperties: false
      required: [code, displayName, startsOn, endsOn]
      properties:
        code: { type: string, minLength: 1, maxLength: 80 }
        displayName: { type: string, minLength: 1, maxLength: 160 }
        startsOn: { type: string, format: date }
        endsOn: { type: string, format: date }
        enabled: { type: boolean }
    AcademicTermList:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/AcademicTerm' } }
        total: { type: integer, minimum: 0 }
...
    CounselorStudentSummary:
      type: object
      required: [studentId, username, displayName, studentNumber, college, major, cohortYear, className, courseCount, totalCredits, averageScorePercentage, averageMastery, highRiskCourseCount, lastActivityAt]
      properties:
        studentId: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        studentNumber: { type: string }
        college: { type: string }
        major: { type: string }
        cohortYear: { type: integer }
        className: { type: string }
        courseCount: { type: integer, minimum: 0 }
        totalCredits: { type: number, minimum: 0 }
        averageScorePercentage: { type: [number, 'null'], minimum: 0, maximum: 100 }
        averageMastery: { type: [number, 'null'], minimum: 0, maximum: 1 }
        highRiskCourseCount: { type: integer, minimum: 0 }
        lastActivityAt: { type: [string, 'null'], format: date-time }
    CounselorStudentPage:
      type: object
      required: [items, total, page, size]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/CounselorStudentSummary' } }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
...
    CounselorStudentOverview:
      type: object
      required: [student, courses]
      properties:
        student: { $ref: '#/components/schemas/CounselorStudentSummary' }
        courses: { type: array, items: { $ref: '#/components/schemas/CounselorCourseStatistics' } }
...
    CounselorClassComparisonResult:
      type: object
      additionalProperties: false
      required: [classes, period, riskTrend]
      properties:
        classes: { type: array, items: { $ref: '#/components/schemas/CounselorClassComparison' } }
        period: { type: string, enum: [7D, 30D, TERM] }
        riskTrend:
          type: array
          items: { $ref: '#/components/schemas/ClassRiskTrendPoint' }
...
    FieldViolation:
      type: object
      additionalProperties: false
      required: [field, message]
      properties:
        field:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
    ProblemDetail:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, traceId, violations]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
          minLength: 1
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 1
        traceId:
          type: string
          minLength: 1
        violations:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'